# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Time Tracking API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Client {
  id: number;
  name: string;
  active: boolean;
  /** @nullable */
  notes: string | null;
  createdAt: string;
}

export interface CreateClientBody {
  name: string;
  active?: boolean;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateClientBody {
  name?: string;
  active?: boolean;
  /** @nullable */
  notes?: string | null;
}

export interface Project {
  id: number;
  clientId: number;
  /** @nullable */
  clientName: string | null;
  name: string;
  /** @nullable */
  code: string | null;
  active: boolean;
  isBillable: boolean;
  /** @nullable */
  budgetHours: number | null;
  /** @nullable */
  startDate: string | null;
  /** @nullable */
  endDate: string | null;
  createdAt: string;
}

export interface CreateProjectBody {
  clientId: number;
  name: string;
  /** @nullable */
  code?: string | null;
  active?: boolean;
  isBillable?: boolean;
  /** @nullable */
  budgetHours?: number | null;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
}

export interface UpdateProjectBody {
  clientId?: number;
  name?: string;
  /** @nullable */
  code?: string | null;
  active?: boolean;
  isBillable?: boolean;
  /** @nullable */
  budgetHours?: number | null;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
}

export interface Employee {
  id: number;
  name: string;
  /** @nullable */
  email: string | null;
  weeklyCapacityHours: number;
  workingDaysMask: number[];
  /** @nullable */
  holidayCalendarCode: string | null;
  personalAccessToken: string;
  active: boolean;
  createdAt: string;
}

export interface EmployeePublic {
  id: number;
  name: string;
}

export interface CreateEmployeeBody {
  name: string;
  /** @nullable */
  email?: string | null;
  weeklyCapacityHours: number;
  workingDaysMask: number[];
  /** @nullable */
  holidayCalendarCode?: string | null;
  active?: boolean;
  pin: string;
}

export interface UpdateEmployeeBody {
  name?: string;
  /** @nullable */
  email?: string | null;
  weeklyCapacityHours?: number;
  workingDaysMask?: number[];
  /** @nullable */
  holidayCalendarCode?: string | null;
  active?: boolean;
}

export interface ResetPinBody {
  pin: string;
}

export interface HolidayCalendar {
  id: number;
  code: string;
  name: string;
  createdAt: string;
}

export interface CreateHolidayCalendarBody {
  code: string;
  name: string;
}

export interface Holiday {
  id: number;
  calendarId: number;
  date: string;
  name: string;
}

export interface CreateHolidayBody {
  date: string;
  name: string;
}

export interface TimeEntry {
  id: number;
  employeeId: number;
  projectId: number;
  /** @nullable */
  projectName: string | null;
  /** @nullable */
  clientName: string | null;
  /** @nullable */
  isBillable: boolean | null;
  entryDate: string;
  hours: number;
  /** @nullable */
  note: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateTimeEntryBody {
  employeeId: number;
  projectId: number;
  entryDate: string;
  hours: number;
  /** @nullable */
  note?: string | null;
}

export interface UpdateTimeEntryBody {
  projectId?: number;
  entryDate?: string;
  hours?: number;
  /** @nullable */
  note?: string | null;
}

export interface BulkTimeEntryItem {
  employeeId: number;
  projectId: number;
  entryDate: string;
  hours: number;
  /** @nullable */
  note?: string | null;
}

export interface BulkUpsertTimeEntriesBody {
  entries: BulkTimeEntryItem[];
}

export interface VerifyPinBody {
  token: string;
  pin: string;
}

export interface UtilizationRow {
  employeeId: number;
  employeeName: string;
  availableHours: number;
  billableHours: number;
  nonBillableHours: number;
  totalBookedHours: number;
  billableUtilization: number;
  overallUtilization: number;
}

export interface ProjectReportRow {
  projectId: number;
  projectName: string;
  clientName: string;
  isBillable: boolean;
  totalHours: number;
  billableHours: number;
  nonBillableHours: number;
}

export interface ClientReportRow {
  clientId: number;
  clientName: string;
  totalHours: number;
  billableHours: number;
  nonBillableHours: number;
}

export interface EmployeeWeekSummary {
  employeeId: number;
  employeeName: string;
  availableHours: number;
  bookedHours: number;
  billableHours: number;
  utilization: number;
}

export interface DashboardSummary {
  weekStartDate: string;
  weekEndDate: string;
  totalBookedHours: number;
  billableBookedHours: number;
  employeeSummaries: EmployeeWeekSummary[];
}

export type ListClientsParams = {
  includeInactive?: boolean;
};

export type ListProjectsParams = {
  includeInactive?: boolean;
  clientId?: number;
};

export type ListEmployeesParams = {
  includeInactive?: boolean;
};

export type ListHolidaysParams = {
  year?: number;
};

export type ListTimeEntriesParams = {
  employeeId?: number;
  projectId?: number;
  startDate?: string;
  endDate?: string;
};

export type GetUtilizationReportParams = {
  startDate: string;
  endDate: string;
  employeeId?: number;
};

export type GetProjectReportParams = {
  startDate: string;
  endDate: string;
};

export type GetClientReportParams = {
  startDate: string;
  endDate: string;
};