# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Time Tracking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      parameters:
        - name: includeInactive
          in: query
          schema:
            type: boolean
        - name: clientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    delete:
      operationId: deleteProject
      tags: [projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── Employees ────────────────────────────────────────────────────────────────
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List employees
      parameters:
        - name: includeInactive
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
    patch:
      operationId: updateEmployee
      tags: [employees]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmployeeBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
    delete:
      operationId: deleteEmployee
      tags: [employees]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /employees/{id}/reset-pin:
    post:
      operationId: resetEmployeePin
      tags: [employees]
      summary: Reset employee PIN and generate new token
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPinBody"
      responses:
        "200":
          description: Pin reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"

  # ── Holiday Calendars & Holidays ────────────────────────────────────────────
  /holiday-calendars:
    get:
      operationId: listHolidayCalendars
      tags: [holidays]
      responses:
        "200":
          description: List of calendars
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HolidayCalendar"
    post:
      operationId: createHolidayCalendar
      tags: [holidays]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHolidayCalendarBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HolidayCalendar"
  /holiday-calendars/{id}/holidays:
    get:
      operationId: listHolidays
      tags: [holidays]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of holidays
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Holiday"
    post:
      operationId: createHoliday
      tags: [holidays]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHolidayBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Holiday"
  /holidays/{id}:
    delete:
      operationId: deleteHoliday
      tags: [holidays]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Time Entries ────────────────────────────────────────────────────────────
...
  /auth/employee/verify:
    post:
      operationId: verifyEmployeePin
      tags: [auth]
      summary: Verify employee PIN for personal link access
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPinBody"
      responses:
        "200":
          description: Employee info on success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "401":
          description: Invalid PIN
  /auth/employee/token/{token}:
    get:
      operationId: getEmployeeByToken
      tags: [auth]
      summary: Get employee basic info by token (without PIN — used to show the name on the PIN entry page)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Employee info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeePublic"
        "404":
          description: Not found
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        name:
          type: string
        code:
          type: ["string", "null"]
        active:
          type: boolean
        isBillable:
          type: boolean
        budgetHours:
          type: ["number", "null"]
        startDate:
          type: ["string", "null"]
          format: date
        endDate:
          type: ["string", "null"]
          format: date
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clientId
        - clientName
        - name
        - code
        - active
        - isBillable
        - budgetHours
        - startDate
        - endDate
        - createdAt
    CreateProjectBody:
      type: object
      properties:
        clientId:
          type: integer
        name:
          type: string
        code:
          type: ["string", "null"]
        active:
          type: boolean
        isBillable:
          type: boolean
        budgetHours:
          type: ["number", "null"]
        startDate:
          type: ["string", "null"]
          format: date
        endDate:
          type: ["string", "null"]
          format: date
      required:
        - clientId
        - name
    UpdateProjectBody:
      type: object
      properties:
        clientId:
          type: integer
        name:
          type: string
        code:
          type: ["string", "null"]
        active:
          type: boolean
        isBillable:
          type: boolean
        budgetHours:
          type: ["number", "null"]
        startDate:
          type: ["string", "null"]
          format: date
        endDate:
          type: ["string", "null"]
          format: date

    # ── Employee ──────────────────────────────────────────────────────────────
    Employee:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        weeklyCapacityHours:
          type: number
        workingDaysMask:
          type: array
          items:
            type: integer
        holidayCalendarCode:
          type: ["string", "null"]
        personalAccessToken:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - weeklyCapacityHours
        - workingDaysMask
        - holidayCalendarCode
        - personalAccessToken
        - active
        - createdAt
    EmployeePublic:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
      required:
        - id
        - name
    CreateEmployeeBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: ["string", "null"]
        weeklyCapacityHours:
          type: number
        workingDaysMask:
          type: array
          items:
            type: integer
        holidayCalendarCode:
          type: ["string", "null"]
        active:
          type: boolean
        pin:
          type: string
      required:
        - name
        - weeklyCapacityHours
        - workingDaysMask
        - pin
    UpdateEmployeeBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: ["string", "null"]
        weeklyCapacityHours:
          type: number
        workingDaysMask:
          type: array
          items:
            type: integer
        holidayCalendarCode:
          type: ["string", "null"]
        active:
          type: boolean
    ResetPinBody:
      type: object
      properties:
        pin:
          type: string
      required:
        - pin

    # ── Holiday Calendars ──────────────────────────────────────────────────
    HolidayCalendar:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - code
        - name
        - createdAt
    CreateHolidayCalendarBody:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
      required:
        - code
        - name
    Holiday:
      type: object
      properties:
        id:
          type: integer
        calendarId:
          type: integer
        date:
          type: string
          format: date
        name:
          type: string
      required:
        - id
        - calendarId
        - date
        - name
    CreateHolidayBody:
      type: object
      properties:
        date:
          type: string
          format: date
        name:
          type: string
      required:
        - date
        - name

    # ── Time Entry ──────────────────────────────────────────────────────────
...
    VerifyPinBody:
      type: object
      properties:
        token:
          type: string
        pin:
          type: string
      required:
        - token
        - pin

    # ── Reports ──────────────────────────────────────────────────────────────
...