# lib/api-zod/src/generated/types/createEmployeeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Time Tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateEmployeeBody {
  name: string;
  /** @nullable */
  email?: string | null;
  weeklyCapacityHours: number;
  workingDaysMask: number[];
  /** @nullable */
  holidayCalendarCode?: string | null;
  active?: boolean;
  pin: string;
}