# lib/api-zod/src/generated/types/updateProjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Time Tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProjectBody {
  clientId?: number;
  name?: string;
  /** @nullable */
  code?: string | null;
  active?: boolean;
  isBillable?: boolean;
  /** @nullable */
  budgetHours?: number | null;
  /** @nullable */
  startDate?: Date | null;
  /** @nullable */
  endDate?: Date | null;
}